Fix thread hierarchy
Messages would update the current parent (to be used for subsequent
children) to themselves before linking with their actually parent. A
child cannot be its own parent, so this broke the thread hierarchy.

This waits to update the parent from the recursive call to after the
message have been linked with its actual parent.